v1.2.0 Added firmware checker and fixed flashing error with the vertiq 8108

+++ App/Schmi/binary_file_interface.hpp
@@ -6,7 +6,7 @@
 namespace Schmi {
 
 struct BytesData {
-  uint16_t num_bytes;
+  uint32_t num_bytes;
   uint32_t starting_byte;
 };
